Update 8
Added a new input method so I don't have to type .strip().lower() over and over. Also went through and made sure the input was on a separate line from the question because it looks a little better.  Also added print3 in order to have a print with a space after for the questions. Also cleaned up a couple areas I'd missed earlier so everything should be the same format now. Changed the pause to 0.5 seconds for now to make it faster to test.  I'll probably take it back up once I'm done.