Functions in case of typo
I didn't want to have to re-type the choice every time in the "else"
function, so I created a global variable "cu_te" (current text) that
would hold the current question for the user.  Then I created a function
that could easily change that variable globally from inside another
function.  Then I created a function that would be able to update the
variable and print it at the same time.  Once all that was set, I
created a function that would print "what?" and then use that global
variable to repeat the current question.